+++ src/stk/molecular/topology_graphs/cof/periodic_honeycomb.py
@@ -10,11 +10,7 @@ import warnings
 from ...reactions import GenericReactionFactory
 from .cof import Cof
 from .vertices import _LinearCofVertex, _NonLinearCofVertex
-from ..topology_graph import (
-    Edge,
-    NullOptimizer,
-    PeriodicConstructionResult,
-)
+from ..topology_graph import Edge, NullOptimizer
 from ...periodic_info import PeriodicInfo
 
 
@@ -161,22 +157,6 @@ class PeriodicHoneycomb(Cof):
                 lattice_constants[2]*self._lattice_size[2]*self._scale
             ),
         )
-
-    def construct(self):
-        """
-        Construct a :class:`.ConstructedMolecule`.
-
-        Returns
-        -------
-        :class:`.PeriodicConstructionResult`
-            The data describing the :class:`.ConstructedMolecule`.
-
-        """
-
-        return super().construct()
-
-    def _get_construction_result(self, state):
-        return PeriodicConstructionResult(state)
 
     _lattice_constants = _a, _b, _c = (
         np.array([1., 0., 0.]),
